refactor: add components foilder
