+++ src/Search/Search.css
@@ -4,7 +4,7 @@
     margin: 0;
     border: 2px solid #646cff;
     border-radius: 1rem;
-    width: 100%;
+    max-width: 320px;
     height: 2.5rem;
     padding: 0.5rem 0.75rem;
     padding-left: 2rem;
