+++ src/layouts/CardGrid.tsx
@@ -1,47 +1,25 @@
 import React, { CSSProperties, PropsWithChildren } from 'react';
 
-interface TwoColumnProps {
-  leftWidth?: string;
-  rightWidth?: string;
-  oneColumnBreakPoint: number;
+interface CardGridProps {
+  cardWidth?: string;
+  cardGap?: string;
 }
-type TwoColumnComponent = PropsWithChildren<TwoColumnProps>;
+type CardGridComponent = PropsWithChildren<CardGridProps>;
 
-type TwoColumnState = {
-  isWideScreen: boolean;
-};
-
-const wideStyle: CSSProperties = {
+const gridStyle: CSSProperties = {
   display: 'grid',
-  gap: 10,
+  gap: '0.8rem',
+  gridTemplateColumns: 'repeat(auto-fill, 320px)',
+  justifyContent: 'center',
   width: '100%',
 };
 
-export class CardGrid extends React.Component<TwoColumnComponent, TwoColumnState> {
-  constructor(props: TwoColumnComponent) {
+export class CardGrid extends React.Component<CardGridComponent, CardGridProps> {
+  constructor(props: CardGridProps) {
     super(props);
-    this.state = { isWideScreen: document.body.offsetWidth > this.props.oneColumnBreakPoint };
-    this.handleWideScreen = this.handleWideScreen.bind(this);
   }
-  handleWideScreen() {
-    this.setState({
-      isWideScreen: document.body.offsetWidth > this.props.oneColumnBreakPoint,
-    });
-  }
-  componentDidMount(): void {
-    window.addEventListener('resize', this.handleWideScreen);
-  }
-  componentWillUnmount(): void {
-    window.removeEventListener('resize', this.handleWideScreen);
-  }
+
   render() {
-    const gridStyle: CSSProperties = this.state.isWideScreen
-      ? {
-          gridTemplateColumns: '1fr 2fr',
-        }
-      : {
-          gridTemplateColumns: '1fr',
-        };
-    return <div style={{ ...wideStyle, ...gridStyle }}>{this.props.children}</div>;
+    return <div style={gridStyle}>{this.props.children}</div>;
   }
 }
